refactor: split into separate files
* add auth.rs for code flow
* rename cookies.rs to session.rs
* add pause.rs
